requirements.txt
psycopg2-binary==2.9.1
djangorestframework==3.14.2
django-filter==2.4.0

